pyproject.toml
[tool.poetry]
name = "promptforge"
version = "0.1.0"
description = "AI/ML prompt engineering and Agentic AI starter kit"
authors = ["Chris Tilley <chris@criticaltilt.com>"]
readme = "README.md"
packages = [{ include = "agents" }, { include = "scripts" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.1"
typer = "^0.12.3"
rich = "^13.7.1"
pydantic = "^2.8.2"
httpx = "^0.27.0"
tenacity = "^8.4.2"
pyyaml = "^6.0.1"

# Providers (install one or both)
openai = "^1.40.0"
anthropic = "^0.34.0"
# Optional meta-router for local/other providers
litellm = "^1.45.12"

# Data / eval (optional but useful)
numpy = "^2.0.1"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.0"
jupyter = "^1.0.0"

# Frameworks (optional; keep off critical path)
langchain = "^0.2.10"
llama-index = "^0.10.64"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
ruff = "^0.5.7"
black = "^24.8.0"
mypy = "^1.10.0"
types-requests = "^2.32.0.20240712"

[tool.ruff]
line-length = 100
target-version = "py310"
extend-select = ["I"]  # import sorting
ignore = ["E501"]      # let black wrap long strings

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
exclude = ["^notebooks/"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
